# apps/api/docs/openapi.yaml
components:
  schemas:
...
    ConversationalQueryRequestSchema:
      properties:
        conversationId:
          anyOf:
          - type: string
          - type: 'null'
          description: Existing conversation ID
          title: Conversationid
        k:
          default: 5
          description: Number of context chunks
          maximum: 20.0
          minimum: 1.0
          title: K
          type: integer
        memoryWindow:
          anyOf:
          - maximum: 50.0
            minimum: 1.0
            type: integer
          - type: 'null'
          default: 10
          description: Memory window size
          title: Memorywindow
        question:
          description: The question to ask
          title: Question
          type: string
        tenantId:
          anyOf:
          - type: string
          - type: 'null'
          description: Target tenant ID
          title: Tenantid
      required:
      - question
      title: ConversationalQueryRequestSchema
      type: object
    ConversationalQueryResponseSchema:
      properties:
        answer:
          description: Generated answer
          title: Answer
          type: string
        confidence:
          anyOf:
          - maximum: 1.0
            minimum: 0.0
            type: number
          - type: 'null'
          description: Confidence score
          title: Confidence
        conversationId:
          description: Conversation ID
          title: Conversationid
          type: string
        queryId:
          description: Query ID
          title: Queryid
          type: string
        sources:
          description: Source documents used
          items:
            type: object
          title: Sources
          type: array
      required:
      - answer
      - sources
      - conversationId
      - queryId
      title: ConversationalQueryResponseSchema
      type: object
    ConversationsResponseSchema:
      description: Response model for listing conversations.
      properties:
        conversations:
          default: []
          items: {}
          title: Conversations
          type: array
      title: ConversationsResponseSchema
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    QueryRequestSchema:
      properties:
        k:
          default: 5
          description: Number of context chunks to retrieve for generating the answer
          example: 5
          maximum: 20.0
          minimum: 1.0
          title: K
          type: integer
        question:
          description: The question to ask about your documents
          example: What are the key risks mentioned in our quarterly report?
          title: Question
          type: string
        tenantId:
          anyOf:
          - type: string
          - type: 'null'
          description: Target tenant ID (defaults to user's tenant)
          example: tenant_123
          title: Tenantid
      required:
      - question
      title: QueryRequestSchema
      type: object
    QueryResponseSchema:
      properties:
        answer:
          description: AI-generated answer based on retrieved document context
          example: 'Based on the strategy documents, the main competitive advantages
            are: 1) Advanced AI technology, 2) Strong brand recognition, 3) Extensive
            distribution network...'
          title: Answer
          type: string
        confidence:
          anyOf:
          - maximum: 1.0
            minimum: 0.0
            type: number
          - type: 'null'
          description: Confidence score for the generated answer (0.0 = low, 1.0 =
            high)
          example: 0.85
          title: Confidence
        query_id:
          description: Unique identifier for this query, used for tracking and debugging
          example: query_abc123def456
          title: Query Id
          type: string
        sources:
          description: Source documents and chunks used to generate the answer, including
            relevance scores
          example:
          - content: Our competitive advantages include...
            id: doc_123
            score: 0.92
            title: Q3 Strategy Document
          items:
            type: object
          title: Sources
          type: array
      required:
      - answer
      - sources
      - query_id
      title: QueryResponseSchema
      type: object
...
info:
  contact:
    email: support@livingtwin.com
    name: Living Twin API Support
    url: https://github.com/kpernyer/living-twin-monorep
  description: "\n# Living Twin Strategic Intelligence Platform\n\nA comprehensive\
    \ API for strategic intelligence, competitive analysis, and AI-powered business\
    \ insights.\n\n## Key Capabilities\n\n* **Strategic Intelligence**: SWOT analysis,\
    \ Porter's Five Forces, competitive intelligence\n* **RAG System**: Document ingestion,\
    \ semantic search, conversational AI\n* **Health Monitoring**: Comprehensive system\
    \ health and performance monitoring  \n* **Testing Framework**: End-to-end testing\
    \ for strategic intelligence workflows\n\n## Authentication\n\nAll endpoints require\
    \ Firebase authentication via the `Authorization` header:\n```\nAuthorization:\
    \ Bearer <firebase-jwt-token>\n```\n\nDevelopment environments support bypass\
    \ mode for testing.\n\n## Tenant Isolation\n\nAll data is isolated by tenant.\
    \ Cross-tenant access is controlled by role-based permissions.\n"
  license:
    name: Proprietary
    url: https://livingtwin.com/license
  title: Living Twin API
  version: 2.0.0
openapi: 3.1.0
paths:
...
  /query:
    post:
      description: "Perform a semantic search across your ingested documents and get\
        \ an AI-generated answer.\n    \n    This endpoint uses RAG (Retrieval-Augmented\
        \ Generation) to:\n    1. Search through your document collection using semantic\
        \ similarity\n    2. Retrieve the most relevant content chunks\n    3. Generate\
        \ a comprehensive answer using AI with the retrieved context\n    \n    **Use\
        \ Cases:**\n    - Ask questions about your business documents\n    - Research\
        \ specific topics across your knowledge base\n    - Get insights from reports,\
        \ policies, and documentation\n    \n    **Example Questions:**\n    - \"\
        What are our key strategic priorities for 2024?\"\n    - \"What risks were\
        \ identified in the quarterly review?\"\n    - \"How does our pricing compare\
        \ to competitors?\""
      operationId: query_query_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequestSchema'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponseSchema'
          description: AI-generated answer with source citations and confidence score
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Query Documents
  /query/conversation/query:
    post:
      description: Conversational query endpoint with memory.
      operationId: conversational_query_query_conversation_query_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConversationalQueryRequestSchema'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationalQueryResponseSchema'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Conversational Query
  /query/conversations:
    get:
      description: List user's conversations.
      operationId: list_conversations_query_conversations_get
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationsResponseSchema'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Conversations
  /query/conversations/{conversation_id}:
    delete:
      description: Delete a conversation.
      operationId: delete_conversation_query_conversations__conversation_id__delete
      parameters:
      - in: path
        name: conversation_id
        required: true
        schema:
          title: Conversation Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponseSchema'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Conversation
    get:
      description: Get conversation with full message history.
      operationId: get_conversation_query_conversations__conversation_id__get
      parameters:
      - in: path
        name: conversation_id
        required: true
        schema:
          title: Conversation Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationDetailSchema'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Conversation
  /query/debug/rag:
    post:
      description: Debug RAG pipeline - returns detailed retrieval information.
      operationId: debug_rag_query_debug_rag_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequestSchema'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                title: Response Debug Rag Query Debug Rag Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Debug Rag
  /query/ingest/recent:
    get:
      description: Get recently ingested documents.
      operationId: get_recent_documents_query_ingest_recent_get
      parameters:
      - in: query
        name: tenantId
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenantid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecentDocumentsResponseSchema'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Recent Documents
  /query/ingest/status:
    get:
      description: Get ingest job status by jobId.
      operationId: get_ingest_status_query_ingest_status_get
      parameters:
      - in: query
        name: jobId
        required: true
        schema:
          title: Jobid
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestJobStatusSchema'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Ingest Status
  /query/ingest/text:
    post:
      description: Ingest endpoint - thin HTTP layer delegating to domain service.
      operationId: ingest_query_ingest_text_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestRequestSchema'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/IngestAcceptedResponseSchema'
                - $ref: '#/components/schemas/IngestResponseSchema'
                title: Response Ingest Query Ingest Text Post
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Ingest
  /query/ingest/upload:
    post:
      description: Upload and ingest a document file (PDF, DOCX, TXT, MD).
      operationId: upload_file_query_ingest_upload_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_file_query_ingest_upload_post'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResponseSchema'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Upload File
...
servers:
- description: Production server
  url: https://api.livingtwin.com
- description: Staging server
  url: https://staging-api.livingtwin.com
- description: Development server
  url: http://localhost:8000
...